+++ lib/truffle.rb
@@ -38,6 +38,7 @@ require_relative "truffle/agent/slash_commands"
 require_relative "truffle/compaction"
 require_relative "truffle/ignore"
 require_relative "truffle/skills"
+require_relative "truffle/context_files"
 require_relative "truffle/system_prompt"
 require_relative "truffle/tools"